+++ test/m1.cpp
@@ -196,27 +196,27 @@ struct int_from_noddy
 // Returns the length of s's held string
 int f(simple const& s)
 {
-    return strlen(s.s);
+    return static_cast<int>(strlen(s.s));
 }
 
 int f_mutable_ref(simple& s)
 {
-    return strlen(s.s);
+    return static_cast<int>(strlen(s.s));
 }
 
 int f_mutable_ptr(simple* s)
 {
-    return strlen(s->s);
+    return static_cast<int>(strlen(s->s));
 }
 
 int f_const_ptr(simple const* s)
 {
-    return strlen(s->s);
+    return static_cast<int>(strlen(s->s));
 }
 
 int f2(SimpleObject const& s)
 {
-    return strlen(s.x.s);
+    return static_cast<int>(strlen(s.x.s));
 }
 
 // A trivial passthru function for simple objects
